refacto(store): implementing the code layer StateHandler
